Use OpenMoko ID for fx2 dev board
`2A19` is Numato's USB ID, and hence should only be used on Numato boards. 
`1D50` is OpenMoko's old USB ID and we have been assigned a couple of the
numbers in that range for our usage.

The `DID` value should also be different between the atlys board and the fx2 board.